+++ main/transaction/index.ts
@@ -1,15 +1,11 @@
-import BigNumber from 'bignumber.js'
-import { addHexPrefix, intToHex } from '@ethereumjs/util'
+import { addHexPrefix } from '@ethereumjs/util'
 import { TransactionFactory, TypedTransaction } from '@ethereumjs/tx'
-import { Common } from '@ethereumjs/common'
 
 import { AppVersion, SignerSummary } from '../signers/Signer'
-import { GasFeesSource, TransactionData, typeSupportsBaseFee } from '../../resources/domain/transaction'
+import { TransactionData, typeSupportsBaseFee } from '../../resources/domain/transaction'
 import { isNonZeroHex } from '../../resources/utils'
 import chainConfig from '../chains/config'
 import { TransactionRequest, TxClassification } from '../accounts/types'
-
-import type { Gas } from '../store/state'
 
 const londonHardforkSigners: SignerCompatibilityByVersion = {
   seed: () => true,
@@ -91,65 +87,6 @@ function maxFee(rawTx: TransactionData) {
   return 50 * 1e18
 }
 
-function calculateMaxFeePerGas(maxBaseFee: string, maxPriorityFee: string) {
-  const maxFeePerGas = BigNumber(maxPriorityFee).plus(maxBaseFee).toString(16)
-  return addHexPrefix(maxFeePerGas)
-}
-
-function populate(rawTx: TransactionData, chainConfig: Common, gas: Gas): TransactionData {
-  const txData: TransactionData = { ...rawTx }
-
-  // non-EIP-1559 case
-  if (!chainConfig.isActivatedEIP(1559) || !gas.price.fees) {
-    txData.type = intToHex(chainConfig.isActivatedEIP(2930) ? 1 : 0)
-
-    const useFrameGasPrice = !rawTx.gasPrice || isNaN(parseInt(rawTx.gasPrice, 16))
-    if (useFrameGasPrice) {
-      // no valid dapp-supplied value for gasPrice so we use the Frame-supplied value
-      const gasPrice = BigNumber(gas.price.levels.fast as string).toString(16)
-      txData.gasPrice = addHexPrefix(gasPrice)
-      txData.gasFeesSource = GasFeesSource.Frame
-    }
-
-    return txData
-  }
-
-  // EIP-1559 case
-  txData.type = intToHex(2)
-
-  const useFrameMaxFeePerGas = !rawTx.maxFeePerGas || isNaN(parseInt(rawTx.maxFeePerGas, 16))
-  const useFrameMaxPriorityFeePerGas =
-    !rawTx.maxPriorityFeePerGas || isNaN(parseInt(rawTx.maxPriorityFeePerGas, 16))
-
-  if (!useFrameMaxFeePerGas && !useFrameMaxPriorityFeePerGas) {
-    // return tx unaltered when we are using no Frame-supplied values
-    return txData
-  }
-
-  if (useFrameMaxFeePerGas && useFrameMaxPriorityFeePerGas) {
-    // dapp did not supply a valid value for maxFeePerGas or maxPriorityFeePerGas so we change the source flag
-    txData.gasFeesSource = GasFeesSource.Frame
-  }
-
-  const maxPriorityFee =
-    useFrameMaxPriorityFeePerGas && gas.price.fees.maxPriorityFeePerGas
-      ? gas.price.fees.maxPriorityFeePerGas
-      : (rawTx.maxPriorityFeePerGas as string)
-
-  // if no valid dapp-supplied value for maxFeePerGas we calculate it
-  txData.maxFeePerGas =
-    useFrameMaxFeePerGas && gas.price.fees.maxBaseFeePerGas
-      ? calculateMaxFeePerGas(gas.price.fees.maxBaseFeePerGas, maxPriorityFee)
-      : txData.maxFeePerGas
-
-  // if no valid dapp-supplied value for maxPriorityFeePerGas we use the Frame-supplied value
-  txData.maxPriorityFeePerGas = useFrameMaxPriorityFeePerGas
-    ? addHexPrefix(BigNumber(maxPriorityFee).toString(16))
-    : txData.maxPriorityFeePerGas
-
-  return txData
-}
-
 function hexifySignature({ v, r, s }: Signature) {
   return {
     v: addHexPrefix(v),
@@ -191,4 +128,4 @@ function classifyTransaction({
   return TxClassification.NATIVE_TRANSFER
 }
 
-export { maxFee, populate, sign, signerCompatibility, londonToLegacy, classifyTransaction }
+export { maxFee, sign, signerCompatibility, londonToLegacy, classifyTransaction }
